integration of search to UI - iteration 1

+++ models/packets/responses_client/file_response.go
@@ -3,5 +3,5 @@ package responses_client
 type FileResponse struct {
 	Filename string `json:"filename"`
 	MetaHash string `json:"meta-hash"`
-	Size     int    `json:"size"`
+	Size     uint64 `json:"size"`
 }
